fix: resolve macOS build warnings and add MLX doctor coverage for v0.6.0
Move Windows-only functions (resolution_from_model_filename,
packaged_inventory_compiled_contexts_ready and its helper,
windows_ofx_layout and compiled_contexts_ready variables) inside
-Wunused-function and -Wunused-variable errors on macOS.

Fix ambiguous std::to_string call on file_time_type in ofx_instance.cpp
by casting to long long.

Add two unit tests to test_job_orchestrator covering the macOS branch
of summarize_doctor_report (packaged_for_macos / validated_platforms
model scan): healthy MLX bundle and missing primary pack failure.

+++ src/plugins/ofx/ofx_instance.cpp
@@ -276,8 +276,9 @@ std::uint64_t models_bundle_token(const std::filesystem::path& models_root) {
             continue;
         }
 
-        entries.push_back(entry.path().filename().string() + "|" + std::to_string(size_bytes) +
-                          "|" + std::to_string(write_time.time_since_epoch().count()));
+        entries.push_back(
+            entry.path().filename().string() + "|" + std::to_string(size_bytes) + "|" +
+            std::to_string(static_cast<long long>(write_time.time_since_epoch().count())));
     }
 
     std::sort(entries.begin(), entries.end());
